add support for disabling certain commit status updates
Right now the GhprbNoCommitStatus extension can be used to completely
disable status updates during the build. However, it would be nice to
have finer control so that users can choose for which events they would
like the commit to be updated.

This patch adds support for this by adding detection for the "--none--"
key string in the triggeredStatus, startedStatus, and completedStatus
fields.

For completedStatus, it depends on the result chosen, e.g. we may want
to update the build if it failed, but not if it succeeded.